secpars_app/requirements_minimal.txt
streamlit>=1.36.0
google-generativeai>=0.7.0
openai>=1.40.0
python-dotenv>=1.0.0
pypdf>=4.0.0
docx2txt>=0.8
Pillow>=10.0.0
# Core dependencies without ML conflicts 